moved generate token function
